+++ examples/HelloWorld/Realization/ManagingSystem/Nodes/Analysis/Analysis.py
@@ -16,9 +16,15 @@ except (ValueError, ImportError):
 
 #<!-- cc_include START--!>
 from fractions import Fraction
-from lidarocclusion.masks import BoolLidarMask, ProbLidarMask
-from lidarocclusion.sliding_lidar_masks import sliding_lidar_mask, sliding_prob_lidar_mask
-from lidarocclusion.masks import BoolLidarMask
+try:
+    from .lidarocclusion.masks import BoolLidarMask, ProbLidarMask
+    from .lidarocclusion.sliding_lidar_masks import sliding_lidar_mask, sliding_prob_lidar_mask
+    from .lidarocclusion.masks import BoolLidarMask
+except (ValueError, ImportError):
+    from lidarocclusion.masks import BoolLidarMask, ProbLidarMask
+    from lidarocclusion.sliding_lidar_masks import sliding_lidar_mask, sliding_prob_lidar_mask
+    from lidarocclusion.masks import BoolLidarMask
+    
 from typing import List, Tuple, Dict
 import traceback
 import numpy as np
@@ -123,7 +129,8 @@ class Analysis(Node):
 
         #<!-- cc_code_analyse_scan_data END--!>
 
-        self.publish_event(event_key='anomaly')    # LINK <outport> anomaly
+        # TODO: Put desired publish event inside user code and uncomment!!
+        #self.publish_event(event_key='anomaly')    # LINK <outport> anomaly
 
     def register_callbacks(self):
         self.register_event_callback(event_key='new_data', callback=self.analyse_scan_data)     # LINK <eventTrigger> new_data