+++ HISTORY.rst
@@ -68,10 +68,23 @@ robosapiensIO Preview
 
 * AADLIL metamodel package added
 
-0.3.10 (2024-12-03)
+0.3.11 (2024-12-03)
 ~~~~~~~~~~~~~~~~~~
 
 * Added deployment strategies {native python, virtual enviornment python, docker containerization}
 * Updated rpio CLI ( run, platform, transformations)
 * ADDED physical architecture and deployment to AADLIL
 * robosapiensIO backbone generation (containerized)
+
+0.3.19 (2024-12-05)
+~~~~~~~~~~~~~~~~~~
+
+* Updated the run and transformation command (CLI)
+* Added template files (TESTING)
+* Updated requirements.txt for all nodes
+
+0.3.20 (2024-12-05)
+~~~~~~~~~~~~~~~~~~
+
+* Update swc2main
+* swc2code updated (user todo added)